Fix sign mismatch comparison in log.c

+++ src/common/log.c
@@ -431,7 +431,7 @@ logPre(LogLevel logLevel, unsigned int processId, const char *fileName, const ch
     if (logLevel >= logLevelDebug)
     {
         // Adding padding for debug and trace levels
-        for (unsigned int paddingIdx = 0; paddingIdx < ((logLevel - logLevelDebug + 1) * 4); paddingIdx++)
+        for (unsigned int paddingIdx = 0; paddingIdx < (unsigned int)((logLevel - logLevelDebug + 1) * 4); paddingIdx++)
         {
             logBuffer[result.bufferPos++] = ' ';
             result.indentSize++;
